+++ constants.go
@@ -4,7 +4,12 @@ const (
 	MIME_XML  = "application/xml"
 	MIME_JSON = "application/json"
 
-	HEADER_Accept       = "Accept"
-	HEADER_ContentType  = "Content-Type"
-	HEADER_LastModified = "Last-Modified"
+	HEADER_Accept          = "Accept"
+	HEADER_ContentType     = "Content-Type"
+	HEADER_LastModified    = "Last-Modified"
+	HEADER_AcceptEncoding  = "Accept-Encoding"
+	HEADER_ContentEncoding = "Content-Encoding"
+
+	ENCODING_GZIP = "gzip"
+	ENCODING_ZLIB = "zlib"
 )